COR: CADRAN: Passage de la colonne angle en int pour eviter les soucis de locale
git-svn-id: https://svn.abls-habitat.fr/repo/Watchdog@4278 9f22af27-4b39-4cfe-9f85-4bed9fef4693

+++ trunk/Include/Reseaux_atelier.h
@@ -108,7 +108,7 @@
     gint    bit_controle;                                                                                       /* Ixxx, Cxxx */
     guint   position_x;                                                                          /* en abscisses et ordonnées */
     guint   position_y;
-    gfloat  angle;
+    gint    angle;
     gboolean fleche_left;                                                 /* TRUE si la fleche de tendance doit etre à gauche */
     gint    nb_decimal;                                                                       /* Nombre de decimal a afficher */
   };
